requirements.txt
Django==4.2.16
djangorestframework==3.15.2
twilio==9.3.2
psycopg[binary]==3.2.3
openai==1.54.3
google-generativeai==0.8.3
